--
Former-commit-id: c6f8756ac5716d40b91820eae9e60600e846cc93

+++ x/pypdist.py
@@ -336,7 +336,8 @@ def _main():
         'LICENSE',
         'README.rst',
     ]:
-        shutil.copyfile(fn, os.path.join(build_root, fn))
+        if os.path.exists(fn):
+            shutil.copyfile(fn, os.path.join(build_root, fn))
 
     subprocess.check_call(
         [
